Refactor cell-effect applying logic
